requirements.txt
aiogram==3.13.1
gspread
google-auth
aiohttp